add ci api && code format
