+++ support/auth/src/api.rs
@@ -1,3 +1,3 @@
-pub mod auth_kernel_api;
 pub mod auth_crypto_api;
+pub mod auth_kernel_api;
 pub mod auth_mgr_api;
