+++ support/auth/src/dto.rs
@@ -1,2 +1,2 @@
-pub mod auth_kernel_dto;
 pub mod auth_crypto_dto;
+pub mod auth_kernel_dto;
