+++ support/iam/src/console_interface/api.rs
@@ -1,4 +1,4 @@
-pub mod iam_ci_cert_api;
 pub mod iam_ci_app_api;
-pub mod iam_ci_system_api;
-pub mod iam_ci_tenant_api;
+pub mod iam_ci_cert_api;
+pub mod iam_ci_res_api;
+pub mod iam_ci_role_api;
