+++ support/iam/src/console_interface/api/iam_ci_role_api.rs
@@ -1,0 +1,56 @@
+use crate::basic::dto::iam_app_dto::{IamAppAggAddReq, IamAppAggModifyReq};
+use crate::basic::serv::iam_app_serv::IamAppServ;
+use crate::basic::serv::iam_role_serv::IamRoleServ;
+use crate::iam_constants::{self, RBUM_SCOPE_LEVEL_APP};
+use bios_basic::process::task_processor::TaskProcessor;
+use tardis::web::context_extractor::TardisContextExtractor;
+use tardis::web::poem_openapi;
+use tardis::web::poem_openapi::param::Path;
+use tardis::web::poem_openapi::payload::Json;
+use tardis::web::web_resp::{TardisApiResult, TardisResp, Void};
+
+pub struct IamCiRoleApi;
+
+/// # Interface Console Manage Cert API
+///
+/// Allow Management Of aksk (an authentication method between applications)
+#[poem_openapi::OpenApi(prefix_path = "/ci/role", tag = "bios_basic::ApiTag::Interface")]
+impl IamCiRoleApi {
+    /// Batch add Role Rel Account
+    #[oai(path = "/:id/account/batch/:account_ids", method = "put")]
+    async fn batch_add_rel_account(&self, id: Path<String>, account_ids: Path<String>, ctx: TardisContextExtractor) -> TardisApiResult<Void> {
+        let mut funs = iam_constants::get_tardis_inst();
+        funs.begin().await?;
+        let app_id = IamAppServ::get_id_by_ctx(&ctx.0, &funs)?;
+        let split = account_ids.0.split(',').collect::<Vec<_>>();
+        for s in split {
+            IamAppServ::add_rel_account(&app_id, s, true, &funs, &ctx.0).await?;
+            IamRoleServ::add_rel_account(&id.0, s, Some(RBUM_SCOPE_LEVEL_APP), &funs, &ctx.0).await?;
+        }
+        funs.commit().await?;
+        TardisResp::ok(Void {})
+    }
+
+    /// Batch delete Role Rel Account
+    #[oai(path = "/:id/account/batch/:account_ids", method = "delete")]
+    async fn batch_delete_rel_account(&self, id: Path<String>, account_ids: Path<String>, ctx: TardisContextExtractor) -> TardisApiResult<Void> {
+        let mut funs = iam_constants::get_tardis_inst();
+        funs.begin().await?;
+        let split = account_ids.0.split(',').collect::<Vec<_>>();
+        for s in split {
+            IamRoleServ::delete_rel_account(&id.0, s, Some(RBUM_SCOPE_LEVEL_APP), &funs, &ctx.0).await?;
+        }
+        funs.commit().await?;
+        TardisResp::ok(Void {})
+    }
+
+    /// Delete Role Rel Account
+    #[oai(path = "/:id/account/:account_id", method = "delete")]
+    async fn delete_rel_account(&self, id: Path<String>, account_id: Path<String>, ctx: TardisContextExtractor) -> TardisApiResult<Void> {
+        let mut funs = iam_constants::get_tardis_inst();
+        funs.begin().await?;
+        IamRoleServ::delete_rel_account(&id.0, &account_id.0, Some(RBUM_SCOPE_LEVEL_APP), &funs, &ctx.0).await?;
+        funs.commit().await?;
+        TardisResp::ok(Void {})
+    }
+}
