+++ support/iam/src/iam_initializer.rs
@@ -31,7 +31,7 @@ use crate::basic::serv::iam_role_serv::IamRoleServ;
 use crate::basic::serv::iam_set_serv::IamSetServ;
 use crate::console_app::api::{iam_ca_account_api, iam_ca_app_api, iam_ca_res_api, iam_ca_role_api};
 use crate::console_common::api::{iam_cc_account_api, iam_cc_app_api, iam_cc_org_api, iam_cc_res_api, iam_cc_role_api, iam_cc_system_api, iam_cc_tenant_api};
-use crate::console_interface::api::{iam_ci_cert_api, iam_ci_app_api, iam_ci_system_api, iam_ci_tenant_api};
+use crate::console_interface::api::{iam_ci_app_api, iam_ci_cert_api, iam_ci_res_api, iam_ci_role_api};
 use crate::console_passport::api::{iam_cp_account_api, iam_cp_app_api, iam_cp_cert_api, iam_cp_tenant_api};
 use crate::console_system::api::{
     iam_cs_account_api, iam_cs_account_attr_api, iam_cs_cert_api, iam_cs_org_api, iam_cs_res_api, iam_cs_role_api, iam_cs_spi_data_api, iam_cs_tenant_api,
@@ -109,8 +109,8 @@ async fn init_api(web_server: &TardisWebServer) -> TardisResult<()> {
                     iam_ci_cert_api::IamCiCertManageApi,
                     iam_ci_cert_api::IamCiCertApi,
                     iam_ci_app_api::IamCiAppApi,
-                    iam_ci_system_api::IamCiSystemApi,
-                    iam_ci_tenant_api::IamCiTenantApi,
+                    iam_ci_res_api::IamCiResApi,
+                    iam_ci_role_api::IamCiRoleApi,
                 ),
             ),
         )
